Reorganize evaluators module
