+++ .vscode/settings.json
@@ -1,1 +1,3 @@
-{}
+{
+    "testing.automaticallyOpenPeekView": "never"
+}
