+++ src/lib/lang/evaluators/order1/cards.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { evalProgram } from ".";
+import { evalProgram } from "..";
 
 describe( "evalCards", () => {
